refactor: file for numpad, clear, and backspace
