+++ app/src/espresso/kotlin/xyz/lbres/trickcalculator/testutils/espressoHelpers.kt
@@ -5,6 +5,7 @@ import android.view.View
 import androidx.test.espresso.intent.Intents.getIntents
 import androidx.test.espresso.matcher.ViewMatchers.withText
 import org.hamcrest.Matcher
+import org.hamcrest.Matchers.anyOf
 import org.hamcrest.Matchers.not
 import org.junit.Assert.assertEquals
 import java.lang.AssertionError
@@ -35,3 +36,8 @@ fun withEmptyString(): Matcher<View> = withText("")
  * [Matcher] that checks if text of TextView is a non-empty string
  */
 fun withNonEmptyString(): Matcher<View> = not(withText(""))
+
+fun withAnyText(options: Iterable<String>): Matcher<View> {
+    val matchers = options.map { withText(it) }.toMutableList()
+    return anyOf(matchers)
+}
